pyproject.toml
[project]
name = "langchain-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.1",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.14",
    "langsmith>=0.4.59",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.15",
]
